Updated get_pts to respect current filters
This commit updates the get_pts() function to manually filter the
points for any existing filters on the selected spatial dimension. One
of the caveates with crossfilters is that it does not remove
categories for filtered dimension if those filters apply to the same
dimension (see crossfilter gotchs faw).  This resulted in all of lake
appearing on the maps when the spatial unit was lake and lakes where
filtered.  This commit ensure that those filters are reflected
correctly on the map.