Updated the stats panel table to use slice label lookup objects
This commit updates the stats panel table to use the longer, more
informative labels from the slice label lookup array. It also updates
the selected pie chart table to use the same, and include the svg
color rectangles.  The table title of the selected pie chart now
includes the pie chart label instead of the slug.