Added active flag to species, strains, and, strain raw
This commit adds an active flag to species, strains, and, strain raw
so that the choices made available in the api can be filtered and
controlled.  The active flag can be changed in the admin, and is also
presented on the lookup table page as a yen/no column.  This commit
also includes a migration that update the active flag using
python. Any value that has not been used in a stocking event since
2000 was updated to inactive.  All previously existing tests continue
to pass.