Added active flag to hatcheries model
This commit adds an active/inactive flag to the hatchery model so that
the values available in the submission template can be controlled.
The active status of a hatchery is presented in the lookups table
template,  and the admin list and can be toggled from within the admin
interface.  The api endpoint for hatchery now accepts and optional
active argument. All previously passing tests continue to pass.