Updated validation to work with the current xls template
This commit includes a number of small changes to ensure that the xls
upload and validation work as they should with the new xls
template. Specifically it adds functionality to find initial values in
the choice labels, in addition to the choice values (so the
spreadsheet can have more meaningful entries) and uses raw strain
rather than strain.  It also transforms the stocking site fields to
title case so folks don't shout TITLE PLACE NAMES.  It also moves the
spreadsheet data template name to the settings file just in case we
change that in the future.  All previously passing tests continue to
pass.