Added clean method to StrainRaw model
This commit adds a clean method to the StrainRaw model to ensure that
the associated species and the species associated with the strain are
consistent.  A unit test ensure that this functionality works as
expected. Additionally the admin interface is updated to limit the
choice of strains to only those associated with the current
species. All previously passing tests continue to pass.