Improved xlsx upload validation
This commit fixes several issues with the upload validation. It
updates the initial value to match either the value or the label (both
are used in excel). Stocking method and life stage are matched without
regard to case.  Additionally stocking sites are always transformed to
proper case, and only active species, strains, and hatcheries are
included in the available choices for each input.  All previously
passing tests continue to pass.