Added comment to event upload and tweaked event clips and marks
This commit updates the event upload by adding a comment field to both
the model and the form. It also includes and update the clips and
marks attributes associated with stocking events to ensure that they
render appropriately when listed in templates.  Some tests were
updated to accommodate these changes. All previously passing tests
continue to pass.