Fixed a large number of issues identified by USFWS team
This commit addresses a fairly large number of small issues that had
been identified by the USWFS team.  Updates include: a warning message
on the project detail form that the event coordinate can be set by
either clicking on the map or entering the lat and lon in the boxes on
the left, a button was added to upload more events to the upload
detail page, tweaks to optimize queries associated with upload list
and upload details page, the loading spinner was still spinning if you
returned to the find event form using the back arrow, and finally, the
map zoom for both found events and cwt events zoom to a more
reasonable limit if only one point is returned.