Wrapped the event upload in a cursor that suspends triggers
this commit wrapped the event upload transaction in a cursor that
temporarily disables the cwt_reused trigger.  Once the transaction has
been completed, the trigger is restored. There is a substantial
improvement in upload speed if the transaction is suspended. The
associated tests were updated with transaction=True. All previously
passing tests continue to pass.