Updated admin search criteria and make figure width dynamic
This commit improved the search criteria in the django admin panel for
several models that were missing important fields that people are
likely to search by - specifically abbrev for hatcheries and strain,
and agency stocking id number.  It also update the template and
associated js source file for found stocking events  to dynamically
calculate and set figure size depending on the size the screen of
viewing device.  All previously passing tests continue to pass.