Removed 'under-development' banner and marked some tests with xfail
This commit prepares the application for final deployment. It removes
red warning banner about the application being under development (as
is should now be considered in production). It also removes dead links
to elements that have not been created and fixes a couple of errors that
where preventing sequential tags start and end from displaying in the
event detail page. Additionally, this commit include updates to the
scripts that were used to migrate the data a final time and patch the
Ontario micro-mark and sequential tag records.  All previously passing
tests continue to pass.