Update centroids to reflect new geom for modified grids
The geometries for several grids were updated to reflect the reality
on the water - grids in Lake Erie where clipped to the shoreline,
while others were updated to accommodated isolated pockets.  This commit
also includes small updates to the scripts used to populate the
postgres database from ms access as well as patch the Ontario-Huron
cwts.  This should be the last time we need those scripts.