Added event_tag_numbers to cwt api and increased tag_no size
This commit adds a field event_tag_numbers to the cwt api serializer
and views to include the actual string of cwts reported with each
stocking event and make it explicit that some event attributes will be
duplicated because of the many-to-many relationship. It also includes
two migrations that increase the maximum size of tag_no to 150 and
updates it to include the concatenated string of cwt numbers. The
remaining updates in this commit are changes to tests that were
failing because of the above changes.