Updates to quick search, xls location, download limit
This commit fixes a number of small bugs that have been identified by
users. The quick search icon has been updated to respond and submit to
mouse clicks.  The form will only be submitted if the input is not
empty.  The primary and secondary stocking locations were switched in
the xlsx download, and a limit was placed on the visualize and
download buttons.  They are now only visible if the number of selected
records is below the the max_event_threshold specified in the settings
file. All previously passing tests continue to pass.