Added stock_id__in filter and padding to event detail form
This commit adds an additional attribute to stocking event form to
find specific stocking events.  It also updated the stocking event
form by adding some padding to the bottom of the form to ensure that
the submit, cancel, and reset buttons are always visible, and changed
the message about clicking on the map to be an actual sentence. All
existing tests continue to pass.