Fixed first year last year bug in FindEvents form
This commit updates the clean method of the FindEvents form to
correctly return the whole dictionary when first and last year are
provided. Originally, they were replacing all other arguments if they
were provided too.  A test was also added to ensure that this is
captured in the future if it happens again. All tests continue to pass.