Added StringAgg(phys_chem_marks)to xls api endpoint
This commit updates the xls api endpoint for stocking events to
aggregate multiple marks associated with an event into a semi-colon
separated string.  In rare instances were multiple marks had been
associated with an event, the event was being repeated once for each
mark. This commit ensure that does not happen. A test has been added
to ensure that response contains a single entry and that the entry has
the expected values for physical-chemical-marks.  All previously
passing tests continue to pass.