Added to fields to Lake for lat-lon bounds
This commit adds four new fields to the Lake model to accommodate the
min and max lat and lon (which are likely to be different than the
extent of the shoreline geometry) and makes them easy to edit and
update in the admin.  The values are used to center any lake-specific
maps and for event detail validation.  Migration files are included to
add the fields and then populate values based on the extend of
stocking events associated with each lake. Tests are included to
ensure that the min and max are always properly entered and that the
coordinate_limits method works was expected.  All previously passing
tests continue to pass.