Moved api documentation and increased stocking event comment size
This commit moves the api docs to ~/api/v1 to make it clear
which api version they are associated with, and increased the size of
the comment field to 5000 characters to ensure that we are able to
capture pretty well anything associated with a stocking event.  All
tests continue to pass.