Updated js for basin-wide pie-charts to use strain slug
The main map has an issues with strains using arbitrary colors rather
than the assign colors - the issue was cause by the api returning the
strain label rather than the slug. A snippet of code was added to the
js to replace the label with the slug on the strain attribute of each
stocking events. All of the pie charts and associated widgets work as
expected.