Updated found cwt events form and template to match found events
This commit includes a small number of changes to the find cwt event
form and associated results table. The return to filter form link was
removed from results page which was returning to an empty form,
instead of the form populated with current filter criteria. The
JavaScript in the form was also updated to ensure that spinner wasn't
showing if the user returned. The behavior of the form is now the
same as the stocking events form and results template. All tests
continue to pass.