Added STP logo, clarified Yreq and relabeled weight to total weight
This commit add a GLFC-Science Transfer Program logo to the footer
block so that it appears on every page of the application. It also
updates the event detail page to stay "Total Weight (kg)" to match the
excel spreadsheet and update the description of  yearling equivalents
to include the rule of 10 for those species and life stages without
published values. All previously passing tests continue to pass.