Added google analytics, increased western bound, c-w-t detail page
This commit includes a number of small changes. Specifically it add a
snippet the the head of the base template to collect data for google
analytics, updates the title attribute of the cwt detail page to
correctly display the cwt number, and extends the westernmost limit of
the bounds for xlsx upload. All tests continue to pass.