Added Geom and grid10 list update from file to spatial admin models
This commit adds functionality to update geometries associated with
Lakes, Jurisdictions, Management Units and Grids from a text file that
contains either Well-Known-Text of geojson. The list of grids
associated with a management unit can now be updated by uploading a
csv of new grid slugs that should be associated with the management
unit in question. All of these functions were tested manually, but no
automated tests are included. All tests continue to pass.