Custom xls render in include metadata tab and event save tag_no
This commit includes a custom xls renderer that inherits from the base
XLS renderer and a meta tag to the downloaded spread that includes the
data and url with query parameters that were used to generate it. The
new renderer is used for both stocking and cwt events. This commit
also adds a method to the stocking event save method to update the
tag_no field with a sorted, semi-colon separated list of cwts
associated with this event.  tag_no us used a short cut to make
serializing more efficient. All tests continue to pass.