Updated the event upload to use set() for many-to-many fields
the original stocking event upload used add to create new stocking
events. This is appropriate when we are only creating new items, but
will potentially create inappropriate assocaitions when those items
are updated. This commit changes those to set() so that the
many-to-many relationships are limited to those in the current
upload.  All tests continue to pass.