Fixed admin strain bug, added label to some model fields
this commit fixes the raw strain bug that had been identified in the
django admin. The admin interface has been updated so that species is
not an available option when creating a new raw strain, instead the
species is populated from the strain when the object is saved. this
commit also adds some basic labels to strain and raw strain to help
distinguish them, and the list of available strains is now sorted by
strain name (rather than by species and then strain). All tests
continue to pass.