Extended map bounds in find events and find cwt events
This commit updates the spatial extents of the global leaflet maps and
reset the zoom to fit the bounds of all 5 lakes combined by
default. This will allow users to explore pretty far away from the
Great Lakes shoreline if they wish, but still constrains them to a
reasonable geography. All tests continue to pass.