BUGFIX 94 - Upload list timestamp not sorting properly
There was a custom parser on the timestamp for the upload list table,
but it wasn't working properly.  The solution was to remove the custom
parser and add a data-text attribute to the timestamp field that
contains the time stamp formatted as number. No custom parser
required.  All tests continue to pass.