Ticket 92 - Filtered CWT make points more obvious
This commit updates the javascript file that provides the
interactivity with the filtered cwt view to make points easier to
find.  The size of the points have been increased slightly and
re-factored into variables that can be more easily reset in the future.
The stroke width of the selected points has also been increased to 3.
Additionally, some event handles were added to the table row or point
on the map will select and highlight that event in both places.  All
tests continue to pass.