Ticket 97 - Error on upload looses comments
This commit fixes the bug in the xlsx upload functionality were any
associated comment would be lost if the upload was not valid. Any
comments are now preserved. The existing tests that checked for
specific failing instances now also check that the submitted comment
is included in the returned response. All tests continue to pass.