Updated Strain admin to list raw strains. Checkbox to lookup tables.
This commit updates the lookup table template to includes a check box
on the strain and hatchery tables so that uses toggle all or active
entries. The commit also includes an update to the Strain admin detail
page to include a list of all of associated sub-strains. All tests
continue to pass.