Replaced Condition with Stocking_Mortality
this is a big commit and renames the Condition model to
Stocking_Mortality. It also changes the attribute of that model from
'condition to 'value'.  All of the occurrence of either Condition or
condition have been replaced, and all tests updated accordingly. All
tests continue to pass, although the JavaScript components need to be
checked and updated too.