Added LookupDescription model and admin so they can be edited
This commit a lookup description model to capture the information
about each model - these descriptions can be edited from the admin
page and are presented on the Lookup tables html page. A place holder
message is returned if the description cannot be found. Tests for both
the model and template are included. All tests continue to pass.