Renamed fk field in StockingEvent to strain_alias
This commit changes the name of the foreign key field in the Stocking
Event model from strain_raw to strain_alias. This includes a migration
that make the change and updates the field name, as well as the
cwt-reuse trigger. It also updates every instance in the code base
where the field on that model was referenced.  All of the tests
continue to pass.